+++ CHANGELOG.md
@@ -40,9 +40,9 @@
 - 🟥 upgraded to `strata` v0.2.0 (see [strata changelog](https://github.com/e280/strata/blob/main/CHANGELOG.md)
   - 🟥 using new signals integration
   - 🍏 added `use` hooks:
-    - `use.derive(f)`
+    - `use.derived(f)`
     - `use.lazy(f)`
-    - `use.signal.derive(f)`
+    - `use.signal.derived(f)`
     - `use.signal.lazy(f)`
 - 🟥 replace `$` multitool with new `dom` tool
     - 🟥 rename `Container` type to `Renderable`